backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain-core
langchain-openai
lxml
requests
openai
PyPDF2
